pyproject.toml
[project]
name = "newsfeed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.5",
    "lxml>=6.0.2",
    "simplenote>=2.1.4",
    "textual>=6.6.0",
    "textual-dev>=1.8.0",
]
